Refactor the tests a little
